% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsheet2tbl.R
\name{gsheet2tbl}
\alias{gsheet2tbl}
\title{Download Google sheet as a table}
\usage{
gsheet2tbl(url)
}
\arguments{
\item{url}{the google sheet url}
}
\description{
\code{sheet2tbl} wraps \code{sheet2text} to parse sheets to tables.
If this function produces unexpected results, use a construction like \code{read.csv(text=google_sheet2text(url, format='csv'), stringsAsFactors=FALSE)} to parse manually, with your own options.
The Google sheet must have 'share by link' turned on.
}
\seealso{
Other gsheet.functions: \code{\link{gsheet2text}}
}

