% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sheet2text.R
\name{sheet2text}
\alias{sheet2text}
\title{Download Google sheet as text.}
\usage{
sheet2text(url, format = "csv")
}
\arguments{
\item{url}{the google sheet url}

\item{format}{controls the column separator used. \code{csv} or \code{tsv}}
}
\description{
Downloads a Google sheet as text, using just the url.
The Google sheet must have 'share by link' turned on.
}
\seealso{
Other gsheet.functions: \code{\link{sheet2tbl}}
}

