% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsheet2text.R
\name{gsheet2text}
\alias{gsheet2text}
\title{Download Google sheet as text.}
\usage{
gsheet2text(url, format = "csv")
}
\arguments{
\item{url}{the google sheet url}

\item{format}{controls the column separator used. \code{csv} or \code{tsv}}
}
\description{
Downloads a Google sheet as text, using just the url.
The Google sheet must have 'share by link' turned on.
}
\examples{
\dontrun{
url <- 'docs.google.com/spreadsheets/d/1I9mJsS5QnXF2TNNntTy-HrcdHmIF9wJ8ONYvEJTXSNo'
a <- gsheet2text(url)
b <- read.csv(a)
}
}
\seealso{
\code{\link{gsheet2text}} to download as a table
}

